feat(FormBuilderPage): extract BuilderDesign container into own dir
