--- frontend/src/features/admin-form-builder/components/BuilderContent/index.ts
@@ -1,1 +1,0 @@
-export * from './BuilderContent'
