+++ frontend/src/features/admin-form-builder/components/BuilderDesign/index.ts
@@ -1,0 +1,1 @@
+export { BuilderDesignContainer as default } from './BuilderDesignContainer'
