+++ frontend/src/features/admin-form-builder/components/BuilderDesign/types.ts
@@ -1,0 +1,9 @@
+export type DndPlaceholderProps =
+  | {
+      droppableId: string
+      clientHeight: number
+      clientWidth: string
+      clientY: number
+      clientX: number
+    }
+  | Record<string, never>
